back/requirements.txt
Flask
flask-cors
pillow
pytesseract
google-generativeai
werkzeug
numpy
joblib
tensorflow
Flask
Flask-Cors
numpy
joblib
tensorflow
requests
schedule
pillow
PyMuPDF
pytesseract
google-generativeai
gpt4all

